Moved type synonyms to package.scala, was causing conflict
