--- src/main/scala/algebra/structures/TypeSynonyms.scala
@@ -1,6 +1,0 @@
-package org.torsteinv.zetatypes.algebra
-
-package object structures {
-    type Rational = Fraction[Integer]
-    type Complex  = ComplexNumber[Rational]
-}
