+++ src/main/scala/algebra/structures/package.scala
@@ -3,4 +3,7 @@ package org.torsteinv.zetatypes.algebra
 /** 
  *  
  */
-package object structures {}
+package object structures {
+    type Rational = Fraction[Integer]
+    type Complex  = ComplexNumber[Rational]
+}
