--- src/main/scala/sequences/TypeSynonyms.scala
@@ -1,8 +1,0 @@
-package org.torsteinv.zetatypes
-
-import org.torsteinv.zetatypes.algebra.structures.{Complex}
-    
-package object sequences {
-    type ComplexSequence = AlgebraicSequence[Complex]
-    type SpecialComplexSequence = SpecialAlgebraicSequence[Complex]
-}
