+++ src/main/scala/sequences/package.scala
@@ -1,6 +1,11 @@
 package org.torsteinv.zetatypes
 
+import org.torsteinv.zetatypes.algebra.structures.{Complex}
+
 /** 
  *  
  */
-package object sequences {}
+package object sequences {
+    type ComplexSequence = AlgebraicSequence[Complex]
+    type SpecialComplexSequence = SpecialAlgebraicSequence[Complex]
+}
